Introducing models
Models are the prototypes of buildings. They will be inserted into a db
(a binary file) and will be used to define actual buildings. E.g. I may
have a model called "FooSoftware Ltd" that is an industrial model with
certain set of attributes (capacity, power required...). I can use this
model to instantiate several industrial building that will use the same
model.